Moved the modelbank into the resourcing lib also removed the mutexes in the resourcemanager so a model can load textures. this need to be solved to pass the course

+++ src/core/subsystem/graphics/SSGraphics.h
@@ -30,5 +30,5 @@ private:
 	gfx::GraphicsEngine* m_GraphicsEngine = nullptr;
 	gfx::RenderQueue* m_RenderQueue = nullptr;
 	Camera* m_Camera = nullptr;
-	int m_TestModel;
+	size_t m_TestModel;
 };
